+++ lib/html/form.php
@@ -3,9 +3,6 @@
 // direct access protection
 if(!defined('KIRBY')) die('Direct access is not allowed');
 
-// dependencies
-require_once('html.php');
-
 /**
  * Form
  *